major reorganization of code
