+++ package.json
@@ -14,11 +14,8 @@
     "graphql": "^15.5.1",
     "jwks-rsa": "^1.7.0",
     "jwt-decode": "^3.1.2",
-    "nodemon": "^2.0.7",
     "pg": "^8.5.1",
-    "socket.io": "^2.4.0",
-    "ts-node": "^9.1.1",
-    "typescript": "^4.2.4"
+    "socket.io": "^2.4.0"
   },
   "scripts": {
     "prestart": "tsc --build",
@@ -30,13 +27,20 @@
   "author": "iannotian",
   "devDependencies": {
     "@types/cors": "^2.8.10",
+    "@types/express": "^4.17.12",
     "@types/pg": "^7.14.11",
-    "@types/socket.io": "^2.1.13"
+    "@types/socket.io": "^2.1.13",
+    "nodemon": "^2.0.9",
+    "ts-node": "^10.0.0",
+    "typescript": "^4.3.5"
   },
   "volta": {
     "node": "16.4.0"
   },
   "engines": {
     "node": "16.x"
+  },
+  "repository": {
+    "url": "https://github.com/iannotian/BigBisonChat"
   }
 }
