+++ src/common/index.ts
@@ -1,0 +1,5 @@
+import EventEmitter = require("eventemitter3");
+
+const em = new EventEmitter();
+
+export { em };
