+++ tsconfig.json
@@ -3,9 +3,10 @@
     "target": "ES2020",
     "lib": ["ES2020"],
     "outDir": "dist",
+    "rootDir": "src",
     "module": "CommonJS",
-    "strict": true
+    "strict": true,
+    "skipLibCheck": true
   },
-  "include": ["src/**/*.ts"],
-  "exclude": ["node_modules/"]
+  "include": ["src"]
 }
